feat：自定义代码生成器starter

+++ jc-club-common-starter/easy-gen-code-spring-boot-starter/src/main/java/com/jingdianjichi/core/utils/MySQLToJavaTypeConverter.java
@@ -18,7 +18,7 @@ public class MySQLToJavaTypeConverter {
     private static final Map<String, Item> typeMap = new HashMap<>();
 
     static {
-        typeMap.put("tinyint", new Item(false, "Byte"));
+        typeMap.put("tinyint", new Item(false, "Integer"));
         typeMap.put("smallint", new Item(false, "Short"));
         typeMap.put("mediumint", new Item(false, "Integer"));
         typeMap.put("int", new Item(false, "Integer"));
